# core-common-user-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-user-mgmt
  description: User Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/users
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/users/{userId}:
    get:
      tags:
        - User Accounts
      summary: Get user account by ID
      description: Returns a user account by its ID
      operationId: getUserAccountById
      parameters:
        - name: userId
          in: path
          description: ID of the user account to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    put:
      tags:
        - User Accounts
      summary: Update an existing user account
      description: Updates an existing user account and returns the updated user account
      operationId: updateUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccountDTO'
        required: true
      responses:
        '200':
          description: User account successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    delete:
      tags:
        - User Accounts
      summary: Delete a user account
      description: Deletes a user account by its ID
      operationId: deleteUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User account successfully deleted
        '404':
          description: User account not found
...
  /api/v1/permissions/{permissionId}:
    get:
      tags:
        - Permissions
      summary: Get permission by ID
      description: Returns a permission by its ID
      operationId: getPermissionById
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
    put:
      tags:
        - Permissions
      summary: Update an existing permission
      description: Updates an existing permission and returns the updated permission
      operationId: updatePermission
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionDTO'
        required: true
      responses:
        '200':
          description: Permission successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
    delete:
      tags:
        - Permissions
      summary: Delete a permission
      description: Deletes a permission by its ID
      operationId: deletePermission
      parameters:
        - name: permissionId
          in: path
          description: ID of the permission to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Permission successfully deleted
        '404':
          description: Permission not found
...
  /api/v1/permissions/filter:
    post:
      tags:
        - Permissions
      summary: Get all permissions with filtering
      description: Returns a paginated list of permissions based on filter criteria
      operationId: filterPermissions
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestPermissionDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/permissions:
    post:
      tags:
        - Permissions
      summary: Create a new permission
      description: Creates a new permission and returns the created permission
      operationId: createPermission
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionDTO'
        required: true
      responses:
        '201':
          description: Permission successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDTO'
...
  /api/v1/audit-logs/filter:
    post:
      tags:
        - Audit Logs
      summary: Get all audit logs with filtering
      description: Returns a paginated list of audit logs based on filter criteria
      operationId: filterAuditLogs
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestAuditLogDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
  /api/v1/audit-logs/{auditLogId}:
    get:
      tags:
        - Audit Logs
      summary: Get audit log by ID
      description: Returns an audit log by its ID
      operationId: getAuditLogById
      parameters:
        - name: auditLogId
          in: path
          description: ID of the audit log to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogDTO'
        '404':
          description: Audit log not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogDTO'
  /api/v1/audit-logs/users/{userId}:
    get:
      tags:
        - Audit Logs
      summary: Get audit logs for a user
      description: Returns all audit logs for a specific user
      operationId: getAuditLogsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/audit-logs/resources/{resourceType}/{resourceId}:
    get:
      tags:
        - Audit Logs
      summary: Get audit logs for a resource
      description: Returns all audit logs for a specific resource
      operationId: getAuditLogsByResource
      parameters:
        - name: resourceType
          in: path
          description: Type of the resource
          required: true
          schema:
            type: string
        - name: resourceId
          in: path
          description: ID of the resource
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
  /api/v1/roles/{roleId}/permissions/{permissionId}:
    delete:
      tags:
        - Role Permissions
      summary: Remove permission from role
      description: Removes a permission from a role
      operationId: removePermissionFromRole
      parameters:
        - name: roleId
          in: path
          description: ID of the role
          required: true
          schema:
            type: integer
            format: int64
        - name: permissionId
          in: path
          description: ID of the permission
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Permission successfully removed from role
        '404':
          description: Role permission not found
components:
  schemas:
    UserAccountDTO:
      required:
        - email
        - fullName
        - isActive
        - userType
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        fullName:
          maxLength: 255
          minLength: 1
          type: string
        nickname:
          maxLength: 100
          minLength: 0
          type: string
        email:
          maxLength: 255
          minLength: 0
          type: string
        userType:
          type: string
          enum:
            - EMPLOYEE
            - DISTRIBUTOR
        branchId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        positionId:
          type: string
          format: uuid
        jobTitle:
          maxLength: 255
          minLength: 0
          type: string
        avatarUrl:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://)?[\w\-]+(\.[\w\-]+)+([\w\-\.,@?^=%&:/~\+#]*[\w\-\@?^=%&/~\+#])?$
          type: string
        themePreference:
          type: string
          enum:
            - LIGHT
            - DARK
            - SYSTEM
        languagePreference:
          maxLength: 10
          minLength: 0
          pattern: ^[a-z]{2}(-[A-Z]{2})?$
          type: string
        locale:
          maxLength: 10
          minLength: 0
          pattern: ^[a-z]{2}_[A-Z]{2}$
          type: string
        timezone:
          maxLength: 50
          minLength: 0
          type: string
        contactPhone:
          maxLength: 20
          minLength: 0
          pattern: ^\+?[1-9]\d{1,14}$
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    PermissionDTO:
      required:
        - domain
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          maxLength: 100
          minLength: 1
          pattern: ^[A-Z_]+$
          type: string
        description:
          maxLength: 500
          minLength: 0
          type: string
        domain:
          maxLength: 100
          minLength: 1
          pattern: ^[a-z_]+$
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestPermissionDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/PermissionDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    AuditLogDTO:
      required:
        - action
        - resource
        - resourceId
        - timestamp
        - userAccountId
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        userAccountId:
          type: string
          format: uuid
        action:
          maxLength: 100
          minLength: 1
          type: string
        resource:
          maxLength: 100
          minLength: 1
          type: string
        resourceId:
          maxLength: 255
          minLength: 1
          type: string
        metadata:
          $ref: '#/components/schemas/JsonNode'
        ipAddress:
          maxLength: 45
          minLength: 0
          pattern: ^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$|^(?:[0-9a-fA-F]{1,4}:){7}[0-9a-fA-F]{1,4}$
          type: string
        timestamp:
          type: string
          format: date-time
    FilterRequestAuditLogDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/AuditLogDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    JsonNode:
      type: object
...